requirements.txt
# Core dependencies
crewai>=0.177.0
anthropic>=0.34.0
openai>=1.50.0

# Database connectivity
snowflake-connector-python==3.17.3
python-dotenv==1.0.1

# Data processing
pandas==2.3.2
numpy==2.3.2

# UI framework
gradio>=5.0.0

# Utilities
pydantic>=2.9.0
jsonlines==4.0.0

# Development
pytest>=8.3.0
black>=24.8.0
PyYAML>=6.0
# Add these lines:
openai>=1.50.0
google-generativeai>=0.3.0
litellm>=1.0.0  # Unified interface for multiple providerstqdm>=4.64.0
